+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "ekd"
+name = "sampledkd"
 version = "0.1.0"
 description = "Entropy-guided Knowledge-Distillation (EKD) for LLM reasoning"
 authors = [
@@ -79,4 +79,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["ekd"] 
+packages = ["sampledkd"]
