+++ sampledkd/distill/__init__.py
@@ -1,0 +1,3 @@
+from .trainer import Distiller, kl_from_vocab_samples
+
+__all__ = ["Distiller", "kl_from_vocab_samples"]
