+++ sampledkd/distill/_mixins/amp_oom.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import torch
+
+
+class AmpOomMixin:
+    @staticmethod
+    def _sanitize_logits(x: torch.Tensor, name: str) -> torch.Tensor:
+        """Sanitize logits to prevent NaNs/Infs during training.
+        We might train with lower precision (e.g., fp16), so instability might occur.
+        """
+        # cast to fp32 for stability, clamp, and replace NaN/Inf
+        x = x.float()
+        x = torch.clamp(x, min=-1e4, max=1e4)
+        x = torch.nan_to_num(x, nan=0.0, posinf=1e4, neginf=-1e4)
+        if not torch.isfinite(x).all():
+            print(f"[warn] non-finite after sanitize in {name}")
+        return x
